refactor(p1): extract WebMountInteractionTools.kt — 8 interaction tools (wait/click/type/keys/select/find/scroll/eval)
Single sibling file holding the full WebMount interaction-primitive surface plus the
high-risk wm_eval (mandatoryApproval=true) which keeps its safeEvalPreview helper
adjacent as file-private. All 8 factories take only WebMountDeps; the bridge dispatch
pattern (callBridge + Text result) is identical across them.

Coordinator collapses 8 `private val xxxTool = Tool(...)` defs + the safeEvalPreview
helper into 8 lazy delegators. WebMountPrimitiveTools.kt: 1071 → 674 lines (-37%).

Verification: :app:compileDebugKotlin --offline → BUILD SUCCESSFUL.
